starter_ai_agents/ai_medical_imaging_agent/requirements.txt
streamlit==1.40.2
agno>=2.2.10
Pillow==10.0.0
duckduckgo-search>=6.4.2
google-generativeai==0.8.3